Create Week6
I uploaded my file. I worked on it for hours, and am still not sure why I keep getting an error on the declarative_base(). I went to the website, and it mentioned the 2.0.16 is the current release, so what I entered should work, but I might be missing something.